doc & fix: add swagger base config and fix code smell

+++ src/main/java/tech/gomes/reading/management/dto/book/response/BookStatusCountDTO.java
@@ -4,5 +4,5 @@ import lombok.Builder;
 
 @Builder
 public record BookStatusCountDTO(String status,
-                                 double count) {
+                                 int count) {
 }
